Ensure huge file lengths and offsets are handled correctly.
git-svn-id: svn://svn.daper.net/moc/trunk@2593 910807d9-36e0-0310-a014-e9ea483e2ba4

+++ decoder_plugins/modplug/modplug.c
@@ -118,7 +118,7 @@ static struct modplug_data *make_modplug_data(const char *file) {
     return data;
   }
 
-  ssize_t size = io_file_size(s);
+  off_t size = io_file_size(s);
 
   if (size == -1) {
     decoder_error(&data->error, ERROR_FATAL, 0, "Can't load module: %s", file);
@@ -131,12 +131,12 @@ static struct modplug_data *make_modplug_data(const char *file) {
 //    return data;
 //  }
 
-  data->filedata = (char *)xmalloc(size);
+  data->filedata = (char *)xmalloc((size_t)size);
 
-  io_read(s, data->filedata, size);
+  io_read(s, data->filedata, (size_t)size);
   io_close(s);
 
-  data->modplugfile=ModPlug_Load(data->filedata, size);
+  data->modplugfile=ModPlug_Load(data->filedata, (int)size);
 
   if(data->modplugfile==NULL) {
     free(data->filedata);
